Modificado para memória grande. Preparado para até 64K de RAM. Testado no simulador com memoria da VM de 6K.

+++ TerraControl/src/app/ControlCore.java
@@ -215,7 +215,7 @@ public class ControlCore implements MessageListener
 	 * Send a specific program block
 	 * @param BlockId	Requested data block number
 	 */
-	void sendNewProgBlock(short BlockId){
+	void sendNewProgBlock(int BlockId){
 		controlform.appendControlMsg("ControlCore: sendNewProgBlock()");
 
 		newProgBlockMsg msg = new newProgBlockMsg();
